fixing video product bug

+++ app/Http/Controllers/Admin/AdminProductController.php
@@ -94,7 +94,9 @@ class AdminProductController extends Controller
 
         if ($request->hasFile('image')) {
             if ($product->image && Storage::disk('public')->exists($product->image)) {
-                Storage::disk('public')->delete($product->image);
+                if($product->image != 'images/default_image.png'){
+                    Storage::disk('public')->delete($product->image);
+                }
             }
 
             $imageName = 'images/'.$product->id.'.'.$request->file('image')->extension();
